% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{get_options}
\alias{get_options}
\title{Options available}
\usage{
get_options(field)
}
\arguments{
\item{field}{a string like "region", "year" or "month"}
}
\description{
Returns the available options for a specific field.
}
\examples{
get_options("region")
}
