% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{get_rfr_with_va}
\alias{get_rfr_with_va}
\title{Query EIOPA RFR with volatility adjustment.}
\usage{
get_rfr_with_va(region, year = NULL, month = NULL, format = c("data.frame"))
}
\arguments{
\item{region}{the region to query (examples: "FR". see \code{get_options("region")}).}

\item{year}{the year of the RFR.}

\item{month}{the month of the RFR.}

\item{format}{the format for the output data. Currently the only option is "data.frame".}
}
\value{
A list of class "eiopa_rfr", with components:
\describe{
\item{data}{a data.frame containing the risk-free rate curves.}
\item{metadata}{a data.frame containing metadata about the risk-free rate curves.}
}
}
\description{
This function query and returns the EIOPA risk-free interest rate term structures from an API.

Note: this function is getting the data from an API. Your IP can be temporary or permanently blocked if too many queries are executed.

For optimal performance, we recommend to limit the number of calls to this function.

EIOPA website : https://www.eiopa.europa.eu/tools-and-data/risk-free-interest-rate-term-structures_en
}
\examples{
get_rfr_with_va("FR", 2019, 12)
get_rfr_with_va("BE", 2020, 11)
}
\seealso{
get_rfr, get_rfr_no_va
}
