% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\name{options_rfr_types}
\alias{options_rfr_types}
\title{This returns a list of the available options for the risk-free-rates curves.}
\usage{
options_rfr_types()
}
\description{
This returns a list of the available options for the risk-free-rates curves.
}
