% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{parse_rfr}
\alias{parse_rfr}
\title{Parse the risk free rates API response}
\usage{
parse_rfr(resp, format)
}
\arguments{
\item{resp}{A response from the API (status 200, type/JSON). The response should have a "data" keyword with the value of an array containing the risk free rates.}

\item{format}{One of the output format ("data.frame" is currently the only option).}
}
\value{
A list of class "eiopa_rfr", with at least the following components:
\describe{
\item{data}{a data.frame containing the risk-free rate curves.}
\item{metadata}{a data.frame containing metadata about the risk-free rate curves.}
}
}
\description{
This function is used to parse data received from the API to various formats.
}
