% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{parse_rfr_to_df}
\alias{parse_rfr_to_df}
\title{Parse the risk free rates API response into a dataframe}
\usage{
parse_rfr_to_df(resp)
}
\arguments{
\item{resp}{A response from the API (status 200, type/JSON). The response should have a "data" keyword with the value of an array containing the risk free rates.}
}
\value{
A list of class "eiopa_rfr", with at least the following components:
\describe{
\item{data}{a data.frame containing the risk-free rate curves.}
\item{metadata}{a data.frame containing metadata about the risk-free rate curves.}
}
}
\description{
This function is used to parse data received from the API into data.frame.
}
