% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\name{PATH_GET_OPTIONS}
\alias{PATH_GET_OPTIONS}
\title{Build path to the options endpoint.}
\usage{
PATH_GET_OPTIONS(field)
}
\arguments{
\item{field}{the field name for the endpoint giving the available options}
}
\value{
A string: the path to the options endpoint.
}
\description{
Build path to the options endpoint.
}
