% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{print.eiopa_rfr}
\alias{print.eiopa_rfr}
\title{Print eiopa_rfr object}
\usage{
\method{print}{eiopa_rfr}(x, ...)
}
\arguments{
\item{x}{a response from the API}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An instance of class "eiopa_rfr". It is a list with at least the following components:
\describe{
\item{data}{a data.frame containing the risk-free rate curves.}
\item{metadata}{a data.frame containing metadata about the risk-free rate curves.}
}
}
\description{
Print eiopa_rfr object in a readable format
}
\examples{
resp <- get_rfr_with_va("FR", 2019, 11)
print(resp)
}
