% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{try_GET}
\alias{try_GET}
\title{Wrapper for GET requests}
\usage{
try_GET(x, ...)
}
\arguments{
\item{x}{the url to request}

\item{...}{additional parameters passed to the GET function.}
}
\value{
The result of the GET request.
}
\description{
Wrapper for GET requests
}
