% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_get}
\alias{api_get}
\title{This function is a helper to make get requests to the API.}
\usage{
api_get(path)
}
\arguments{
\item{path}{the path to an endpoint of the API}
}
\value{
A list of class "eiopaR_api". It has at least the following components:
\describe{
\item{content}{the JSON content parsed to a list.}
\item{path}{the path to the requested endpoint.}
\item{response}{the raw response.}
}
}
\description{
This function is a helper to make get requests to the API.
}
