% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{get_rfr}
\alias{get_rfr}
\title{Query EIOPA risk-free rate (RFR).}
\usage{
get_rfr(
  type = options_rfr_types(),
  region,
  year = NULL,
  month = NULL,
  format = c("data.frame")
)
}
\arguments{
\item{type}{\link{string} the type of the risk-free rate, see \code{options_rfr_types()} for the different options (examples: "with_va", "no_va").}

\item{region}{\link{string} the region to query (examples: "FR", "BE")}

\item{year}{\link{integer} the year of the RFR to query.}

\item{month}{\link{integer} the month of the RFR to query.}

\item{format}{\link{string} the format for the output data. Currently the only option is "data.frame".}
}
\description{
This is a generic function to query EIOPA risk-free interest rate term structures.

Note: this function is getting the data from an API. Your IP can be temporary or permanently blocked if too many queries are executed.

For optimal performance of your script, we recommend to call this function only once by session and term structure and to store the data in the environment variables of your session.

EIOPA website : https://www.eiopa.europa.eu/tools-and-data/risk-free-interest-rate-term-structures_en
}
\examples{
get_rfr("with_va", "FR", 2019, 12)
get_rfr("no_va", "FR", c(2016, 2019), 12)
get_rfr("no_va", "FR", 2019, 12)
get_rfr("no_va_shock_up", "BE", 2020, 11)
}
\seealso{
get_rfr_with_va, get_rfr_no_va, get_rfr_with_va_shock_up, get_rfr_with_va_shock_down, get_rfr_no_va_shock_up, get_rfr_no_down
}
