% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{get_rfr_with_va}
\alias{get_rfr_with_va}
\title{Query EIOPA RFR with volatility adjustment.}
\usage{
get_rfr_with_va(region, year = NULL, month = NULL, format = c("data.frame"))
}
\arguments{
\item{region}{the region to query (examples: "FR", "BE")}

\item{year}{the year of the RFR to query.}

\item{month}{the month of the RFR to query.}

\item{format}{the format for the output data. Currently the only option is "data.frame".}
}
\description{
This function query and returns the EIOPA risk-free interest rate term structures from an API.

Note: this function is getting the data from an API. Your IP can be temporary or permanently blocked if too many queries are executed.

For optimal performance of your script, we recommend to call this function only once by session and term structure and to store the data in the environment variables of your session.

EIOPA website : https://www.eiopa.europa.eu/tools-and-data/risk-free-interest-rate-term-structures_en
}
\examples{
get_rfr_with_va("FR", 2019, 12)
get_rfr_with_va("BE", 2020, 11)
}
\seealso{
get_rfr, get_rfr_no_va, get_rfr_with_va_shock_up, get_rfr_with_va_shock_down, get_rfr_no_va_shock_up, get_rfr_no_down
}
