% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{parse_rfr}
\alias{parse_rfr}
\title{Parse the risk free rates API response into dataframe}
\usage{
parse_rfr(resp, format)
}
\arguments{
\item{resp}{(list of list) A response from the API (status 200, type/JSON). The response should have a "data" keyword with value an array containing the risk free rates.}

\item{format}{(string) One of the output format ("data.frame", "data.table", "array").}
}
\description{
This function is used to parse data received from the api to various formats.
}
