% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\name{PATH_GET_RFR}
\alias{PATH_GET_RFR}
\title{Build path to the RFR (risk-free rate) endpoint.}
\usage{
PATH_GET_RFR(type, region, params)
}
\arguments{
\item{type}{one of the types available by the API (see \code{options_rfr_types()})}

\item{region}{one of the regions available by the API (see \code{get_options("region")})}

\item{params}{additional GET parameters}
}
\value{
A string: the path to the RFR endpoint.
}
\description{
Build path to the RFR (risk-free rate) endpoint.
}
