% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiopaR.R
\name{print.eiopa_rfr}
\alias{print.eiopa_rfr}
\title{print risk free rates API response}
\usage{
\method{print}{eiopa_rfr}(x, ...)
}
\arguments{
\item{x}{a response from the api}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Display the response from the API in a readable format.
}
\examples{
resp <- get_rfr_with_va("FR", 2019, 11)
print(resp)
}
